fix(v14,v11a): correct bf16/fp32 dtype handling in model forward
v14_data_dependent: cast the fp32 gate_weight control tensor to the activation dtype before the matmul (model.py:71). Under bf16 autocast the bf16 activation met the fp32 gate_weight and raised 'expected m1 and m2 to have the same dtype'.

v11a_mixed_ops: build EMA position indices in float32 (model.py:60). bf16 cannot represent integers > 256 exactly, so at the training seq_len of 1024 the decay exponents were silently corrupted.

Validated: pytest 58 passed; bf16-autocast CPU smoke 17/17 at seq 64 and 512; single RTX 4090 smoke trains and descends (v14 val_loss 6.30->6.12, v11a 11.74->7.44) with no NaN/instability.

+++ v11a_mixed_ops/model.py
@@ -57,7 +57,9 @@ class HighDecayEMA(nn.Module):
         # Efficient: use log-space cumsum trick
         log_decay = torch.log(decay.clamp(min=1e-6))  # (D,)
         # Weight each position: w[t] = decay^(T-1-t)
-        steps = torch.arange(T, device=x.device, dtype=x.dtype)
+        # fp32 positions: bf16 cannot represent integers > 256 exactly, which
+        # would corrupt the decay exponents at the training seq_len of 1024.
+        steps = torch.arange(T, device=x.device, dtype=torch.float32)
         # weights[t, d] = exp((T-1-t) * log_decay[d])
         weights = torch.exp(steps.flip(0).unsqueeze(-1) * log_decay.unsqueeze(0))  # (T, D)
         # Weighted x, cumsum, then unweight
